refactor: 오타 수정, 에러 처리 개선 및 스토어 기본값 업데이트
## 작업 내용
- SavaButton.tsx → SaveButton.tsx로 파일명 변경 및 Header.tsx의 import 경로 수정
- Galaxy 커뮤니티 리스트에서 에러 바운더리와 reset 로직을 개선하여 에러 처리 강화
- 빈 리스트일 경우 'No data' 메시지 추가 및 디버그 로깅 보강
- useStellarStore, useUserStore의 초기값 업데이트 (galaxy_id와 사용자 인증 상태 기본값 수정)

## 참고 사항
- 파일명 및 경로 불일치 문제를 방지
- 에러 처리 및 기본값 개선으로 코드 안정성과 유지보수성 향상
