+++ src/components/panel/galaxy/community/List.tsx
@@ -1,4 +1,4 @@
-import { Suspense } from 'react';
+import { Suspense, useEffect } from 'react';
 import { useGetGalaxyCommunityList } from '@/hooks/api/useGalaxy';
 import CardItem from './CardItem';
 import {
@@ -11,12 +11,20 @@ import { SkeletonCard } from '@/components/common/Card/SkeletonCard';
 import { ErrorBoundary } from 'react-error-boundary';
 import LoadingIcon from '@/components/common/LoadingIcon';
 import useStellarSystemSelection from '@/hooks/useStellarSystemSelection';
+import { useQueryErrorResetBoundary } from '@tanstack/react-query';
+import type { FallbackProps } from 'react-error-boundary';
 
 const GALAXY_LIST_LIMIT = 3;
 
 export default function List({ sort }: { sort: SortLabel }) {
+  const { reset } = useQueryErrorResetBoundary();
+
   return (
-    <ErrorBoundary FallbackComponent={ErrorComp}>
+    <ErrorBoundary
+      FallbackComponent={ErrorComp}
+      resetKeys={[sort]} // sort 바뀌면 에러 초기화
+      onReset={reset} // React Query 에러 상태도 리셋
+    >
       <Suspense fallback={<LoadingComp />}>
         <ContentComp sort={sort} />
       </Suspense>
@@ -28,6 +36,9 @@ export default function List({ sort }: { sort: SortLabel }) {
 function ContentComp({ sort }: { sort: SortLabel }) {
   const { selectStellar } = useStellarSystemSelection();
 
+  useEffect(() => {
+    console.log('toSortValue(sort) : ', toSortValue(sort));
+  }, [sort]);
   // 갤럭시 리스트 데이터
   const { data, fetchNextPage, hasNextPage, isFetchingNextPage } =
     useGetGalaxyCommunityList({
@@ -37,6 +48,10 @@ function ContentComp({ sort }: { sort: SortLabel }) {
     });
   // 평탄화 된 list 데이터
   const galaxyCommunityList = data?.list ?? [];
+
+  if (galaxyCommunityList.length === 0) {
+    return <div>No data</div>;
+  }
 
   return (
     <div>
@@ -82,6 +97,13 @@ function LoadingComp() {
 }
 
 // 에러 처리
-function ErrorComp() {
-  return <div>Error</div>;
+function ErrorComp({ error, resetErrorBoundary }: FallbackProps) {
+  return (
+    <div className="p-4">
+      <p className="mb-2">에러 발생 : {error.message}</p>
+      <Button color="tertiary" onClick={resetErrorBoundary}>
+        다시 시도
+      </Button>
+    </div>
+  );
 }
